xchainpy/xchainpy_thorchain/requirements.txt
ecdsa==0.16.1
http3==0.6.7
secp256k1==0.13.2
mnemonic==0.19
bech32==1.2.0
binance_chain==1.0.0
hdwallets==0.1.2
xchainpy_client>=0.1.3
xchainpy_crypto>=0.1.5
xchainpy_util>=0.1.5